Outsource spv task details into a different page
