+++ frontend/src/pages/MyStaffs.jsx
@@ -158,6 +158,8 @@ export default function MyStaffs() {
           My Staff's List
         </div>
 
+
+
         {/* Right: two date inputs, each with its own label */}
         <div className="flex space-x-6">
           {/* Start Date */}
@@ -189,6 +191,8 @@ export default function MyStaffs() {
           </div>
         </div>
       </div>
+
+      
 
       {/* Loading state */}
       {loading ? (
